cups-pdf patch for Mac OS X Yosemite permissions
cups-pdf has permission problems with the new sandbox model for cups in
Mac OS X Yosemite 10.10.

This patch captures a hack to make based on the discussion here:
* https://bitbucket.org/codepoet/cups-pdf-for-mac-os-x/issue/55/file-doesnt-print#comment-17921875

The cups-pdf Formula was originally derived from MacPorts.

See more discussion about the Mac OS X Yosemite issues at:
* https://trac.macports.org/attachment/ticket/45006/
* https://github.com/Homebrew/homebrew/issues/40083#issuecomment-105776514

This is not intended to be merged to master but rather available for
discussion and/or testing.